Rename F1 scenario files
