+++ features/F1/test/acceptance.py
@@ -1,11 +1,11 @@
-from .f1s1 import f1s1
-from .f1s2 import f1s2
-from .f1s3 import f1s3
-from .f1s4 import f1s4
-from .f1s5 import f1s5
-from .f1s6 import f1s6
-from .f1s7 import f1s7
-from .f1s8 import f1s8
+from .s1 import f1s1
+from .s2 import f1s2
+from .s3 import f1s3
+from .s4 import f1s4
+from .s5 import f1s5
+from .s6 import f1s6
+from .s7 import f1s7
+from .s8 import f1s8
 
 __all__ = [
     "f1s1",
